refactor(centralServer): Broke surveyResponse.js into multiple files
