+++ packages/central-server/src/apiV2/surveyResponses/index.js
@@ -7,4 +7,7 @@ export { DeleteSurveyResponses } from './DeleteSurveyResponses';
 export { GETSurveyResponses } from './GETSurveyResponses';
 export * from './assertSurveyResponsePermissions';
 export { saveResponsesToDatabase } from './saveResponsesToDatabase';
+export { upsertEntitiesAndOptions } from './upsertEntitiesAndOptions';
+export { validateSurveyResponses } from './validateSurveyResponses';
 export { ResubmitSurveyResponse } from './ResubmitSurveyResponse';
+export { surveyResponse } from './surveyResponse';
