+++ packages/central-server/src/tests/apiV2/utilities/SurveyResponseImporter.test.js
@@ -9,7 +9,7 @@ import sinon from 'sinon';
 
 import { generateId, createModelsStub as baseCreateModelsStub } from '@tupaia/database';
 import { SurveyResponseImporter } from '../../../apiV2/utilities';
-import * as SurveyResponse from '../../../apiV2/surveyResponse';
+import * as SurveyResponse from '../../../apiV2/surveyResponses';
 
 const ENTITY_IDS = {
   1989: generateId(),
@@ -67,13 +67,17 @@ let clock;
 describe('SurveyResponseImporter', () => {
   before(() => {
     clock = sinon.useFakeTimers({ now: TIMESTAMP, toFake: ['Date'] });
+    sinon.stub(SurveyResponse, 'upsertEntitiesAndOptions').callsFake(() => {});
+    sinon.stub(SurveyResponse, 'validateSurveyResponses').callsFake(() => {});
     sinon
-      .stub(SurveyResponse, 'submitResponses')
+      .stub(SurveyResponse, 'saveResponsesToDatabase')
       .callsFake((models, userId, responses) => RESULTS_BY_SURVEY_ID[responses[0].survey_id]);
   });
 
   after(() => {
-    SurveyResponse.submitResponses.restore();
+    SurveyResponse.upsertEntitiesAndOptions.restore();
+    SurveyResponse.validateSurveyResponses.restore();
+    SurveyResponse.saveResponsesToDatabase.restore();
     clock.restore();
   });
 
@@ -89,19 +93,22 @@ describe('SurveyResponseImporter', () => {
     });
 
     beforeEach(() => {
-      SurveyResponse.submitResponses.resetHistory();
+      SurveyResponse.saveResponsesToDatabase.resetHistory();
     });
 
     it('should use the provided user id for the survey submissions', async () => {
       await importer.import(ROWS_BY_SURVEY, USER_ID);
-      expect(SurveyResponse.submitResponses).to.have.been.calledWith(sinon.match.any, USER_ID);
+      expect(SurveyResponse.saveResponsesToDatabase).to.have.been.calledWith(
+        sinon.match.any,
+        USER_ID,
+      );
     });
 
     it('should use the provided response data as survey responses', async () => {
       await importer.import(ROWS_BY_SURVEY, USER_ID);
 
-      expect(SurveyResponse.submitResponses).to.have.been.calledTwice;
-      expect(SurveyResponse.submitResponses).to.have.been.calledWith(
+      expect(SurveyResponse.saveResponsesToDatabase).to.have.been.calledTwice;
+      expect(SurveyResponse.saveResponsesToDatabase).to.have.been.calledWith(
         sinon.match.any,
         sinon.match.any,
         [
@@ -119,7 +126,7 @@ describe('SurveyResponseImporter', () => {
           },
         ],
       );
-      expect(SurveyResponse.submitResponses).to.have.been.calledWith(
+      expect(SurveyResponse.saveResponsesToDatabase).to.have.been.calledWith(
         sinon.match.any,
         sinon.match.any,
         [
